+++ setup.py
@@ -1082,6 +1082,7 @@ def main():
         'include/torch/csrc/jit/codegen/cuda/scheduler/*.h',
         'include/torch/csrc/onnx/*.h',
         'include/torch/csrc/profiler/*.h',
+        'include/torch/csrc/profiler/orchestration/*.h',
         'include/torch/csrc/utils/*.h',
         'include/torch/csrc/tensor/*.h',
         'include/torch/csrc/lazy/backend/*.h',
